Add finialiser function
