+++ pkg/controller/module_controller.go
@@ -31,11 +31,203 @@ type ReconcileModule struct {
 	Scheme *runtime.Scheme
 }
 
+func (r *ReconcileModule) deletionReconcile(moduleInterface interface{}, finalizerInterfaces ...interface{}) error {
+	for _, finalizerInterface := range finalizerInterfaces {
+		switch module := moduleInterface.(type) {
+		case *modulev1alpha1.GCS:
+			for _, fin := range module.GetFinalizers() {
+				instance_split_fin := strings.Split(fin, "_")
+				switch finalizer := finalizerInterface.(type) {
+				case *modulev1alpha1.GCS:
+					if instance_split_fin[0] == "Module" && instance_split_fin[1] == "GCS" && instance_split_fin[2] != module.ObjectMeta.Name {
+						if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: module.ObjectMeta.Namespace}, finalizer); errors.IsNotFound(err) {
+							util.RemoveFinalizer(module, fin)
+							if err := r.Update(context.Background(), module); err != nil {
+								return err
+							}
+						} else {
+							return errors.NewBadRequest("GCS dependency is not met for deletion")
+						}
+					}
+				case *providerv1alpha1.Google:
+					if instance_split_fin[0] == "Provider" && instance_split_fin[1] == "Google" {
+						if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: module.ObjectMeta.Namespace}, finalizer); errors.IsNotFound(err) {
+							util.RemoveFinalizer(module, fin)
+							if err := r.Update(context.Background(), module); err != nil {
+								return err
+							}
+						} else {
+							return errors.NewBadRequest("Google dependency is not met for deletion")
+						}
+					}
+				case *backendv1alpha1.EtcdV3:
+					if instance_split_fin[0] == "Backend" && instance_split_fin[1] == "EtcdV3" {
+						if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: module.ObjectMeta.Namespace}, finalizer); errors.IsNotFound(err) {
+							util.RemoveFinalizer(module, fin)
+							if err := r.Update(context.Background(), module); err != nil {
+								return err
+							}
+						} else {
+							return errors.NewBadRequest("EtcdV3 dependency is not met for deletion")
+						}
+					}
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func (r *ReconcileModule) terraformReconcile(moduleInterface interface{}) error {
+	switch module := moduleInterface.(type) {
+	case *modulev1alpha1.GCS:
+		if err := terraform.TerraformInit(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Init"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if !reflect.DeepEqual("Init", module.Status.Phase) {
+			// Set the data
+			module.Status.State = "Success"
+			module.Status.Phase = "Init"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+		}
+
+		if err := terraform.TerraformNewWorkspace(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Workspace"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if err := terraform.TerraformSelectWorkspace(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Workspace"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if !reflect.DeepEqual("Workspace", module.Status.Phase) {
+			// Set the data
+			module.Status.State = "Success"
+			module.Status.Phase = "Workspace"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+		}
+
+		if err := terraform.TerraformValidate(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Validate"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if !reflect.DeepEqual("Validate", module.Status.Phase) {
+			// Set the data
+			module.Status.State = "Success"
+			module.Status.Phase = "Validate"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+		}
+
+		if err := terraform.TerraformPlan(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Plan"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if !reflect.DeepEqual("Plan", module.Status.Phase) {
+			// Set the data
+			module.Status.State = "Success"
+			module.Status.Phase = "Plan"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+		}
+
+		if err := terraform.TerraformApply(module.ObjectMeta.Namespace); err != nil {
+			// Set the data
+			module.Status.State = "Failure"
+			module.Status.Phase = "Apply"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+			if err := terraform.RemoveFile(module.ObjectMeta.Namespace, module.ObjectMeta.Name); err != nil {
+				return err
+			}
+			return nil
+		}
+
+		if !reflect.DeepEqual("Apply", module.Status.Phase) {
+			// Set the data
+			module.Status.State = "Success"
+			module.Status.Phase = "Apply"
+			// Update the CR with status ready
+			if err := r.Status().Update(context.Background(), module); err != nil {
+				return err
+			}
+		}
+	default:
+		return errors.NewBadRequest("terraform reconcile does not implement on this type")
+	}
+
+	return nil
+}
+
 // +kubebuilder:rbac:groups=batch.my.domain,resources=cronjobs,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=batch.my.domain,resources=cronjobs/status,verbs=get;update;patch
 
 func (r *ReconcileModule) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	module := &modulev1alpha1.GCS{}
+	provider := &providerv1alpha1.Google{}
+	backend := &backendv1alpha1.EtcdV3{}
 
 	if err := r.Get(context.Background(), req.NamespacedName, module); err != nil {
 		if errors.IsNotFound(err) {
@@ -57,48 +249,8 @@ func (r *ReconcileModule) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	}
 
 	if util.IsBeingDeleted(module) {
-		finalizersProvider := &providerv1alpha1.Google{}
-		finalizersModule := &modulev1alpha1.GCS{}
-		finalizersBackend := &backendv1alpha1.EtcdV3{}
-
-		for _, fin := range module.GetFinalizers() {
-			instance_split_fin := strings.Split(fin, "_")
-
-			if instance_split_fin[0] == "Module" && instance_split_fin[1] == "GCS" && instance_split_fin[2] != module.ObjectMeta.Name {
-				if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: req.NamespacedName.Namespace}, finalizersModule); errors.IsNotFound(err) {
-					util.RemoveFinalizer(module, fin)
-
-					if err := r.Update(context.Background(), module); err != nil {
-						return reconcile.Result{}, err
-					}
-				} else {
-					return reconcile.Result{}, errors.NewBadRequest("module dependency is not met for deletion")
-				}
-			}
-
-			if instance_split_fin[0] == "Provider" && instance_split_fin[1] == "Google" {
-				if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: req.NamespacedName.Namespace}, finalizersProvider); errors.IsNotFound(err) {
-					util.RemoveFinalizer(module, fin)
-
-					if err := r.Update(context.Background(), module); err != nil {
-						return reconcile.Result{}, err
-					}
-				} else {
-					return reconcile.Result{}, errors.NewBadRequest("provider dependency is not met for deletion")
-				}
-			}
-
-			if instance_split_fin[0] == "Backend" && instance_split_fin[1] == "EtcdV3" {
-				if err := r.Get(context.Background(), types.NamespacedName{Name: instance_split_fin[2], Namespace: req.NamespacedName.Namespace}, finalizersBackend); errors.IsNotFound(err) {
-					util.RemoveFinalizer(module, fin)
-
-					if err := r.Update(context.Background(), module); err != nil {
-						return reconcile.Result{}, err
-					}
-				} else {
-					return reconcile.Result{}, errors.NewBadRequest("backend dependency is not met for deletion")
-				}
-			}
+		if err := r.deletionReconcile(module, provider, backend); err != nil {
+			return reconcile.Result{}, err
 		}
 
 		if err := terraform.WriteToFile([]byte("{}"), module.ObjectMeta.Namespace, "Module_"+module.Kind+"_"+module.ObjectMeta.Name); err != nil {
